refactor(review): delete dead code from review-engine.ts (#291)
Extract the 2 live functions (filterAnalyzableFiles, runComplexityAnalysis)
into a new analysis.ts module and delete the ~1500-line review-engine.ts
along with its 43 dead-code-only tests.

The remaining 8 exported functions, 1 type, and all internal helpers were
superseded by the plugin architecture (engine.ts + plugins/ + adapters/).
